@@ -12,9 +12,9 @@ import android.content.pm.ResolveInfo;
 import android.media.MediaScannerConnection;
 import android.net.Uri;
 import android.provider.MediaStore;
-import android.support.annotation.VisibleForTesting;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.FileProvider;
+import androidx.annotation.VisibleForTesting;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.FileProvider;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.PluginRegistry;
@@ -186,7 +186,7 @@ public class ImagePickerDelegate
 
   public void chooseVideoFromGallery(MethodCall methodCall, MethodChannel.Result result) {
     if (!setPendingMethodCallAndResult(methodCall, result)) {
-      finishWithAlreadyActiveError();
+      finishWithAlreadyActiveError(result);
       return;
     }
 
@@ -208,7 +208,7 @@ public class ImagePickerDelegate
 
   public void takeVideoWithCamera(MethodCall methodCall, MethodChannel.Result result) {
     if (!setPendingMethodCallAndResult(methodCall, result)) {
-      finishWithAlreadyActiveError();
+      finishWithAlreadyActiveError(result);
       return;
     }
 
@@ -242,7 +242,7 @@ public class ImagePickerDelegate
 
   public void chooseImageFromGallery(MethodCall methodCall, MethodChannel.Result result) {
     if (!setPendingMethodCallAndResult(methodCall, result)) {
-      finishWithAlreadyActiveError();
+      finishWithAlreadyActiveError(result);
       return;
     }
 
@@ -264,7 +264,7 @@ public class ImagePickerDelegate
 
   public void takeImageWithCamera(MethodCall methodCall, MethodChannel.Result result) {
     if (!setPendingMethodCallAndResult(methodCall, result)) {
-      finishWithAlreadyActiveError();
+      finishWithAlreadyActiveError(result);
       return;
     }
 
@@ -482,8 +482,8 @@ public class ImagePickerDelegate
     clearMethodCallAndResult();
   }
 
-  private void finishWithAlreadyActiveError() {
-    finishWithError("already_active", "Image picker is already active");
+  private void finishWithAlreadyActiveError(MethodChannel.Result result) {
+    result.error("already_active", "Image picker is already active", null);
   }
 
   private void finishWithError(String errorCode, String errorMessage) {
