@@ -97,6 +97,10 @@ class AndroidAlarmManager {
   /// alarm fires. If `wakeup` is false (the default), the device will not be
   /// woken up to service the alarm.
   ///
+  /// If `rescheduleOnReboot` is passed as `true`, the alarm will be persisted
+  /// across reboots. If `rescheduleOnReboot` is false (the default), the alarm
+  /// will not be rescheduled after a reboot and will not be executed.
+  ///
   /// Returns a [Future] that resolves to `true` on success and `false` on
   /// failure.
   static Future<bool> oneShot(
@@ -105,6 +109,7 @@ class AndroidAlarmManager {
     dynamic Function() callback, {
     bool exact = false,
     bool wakeup = false,
+    bool rescheduleOnReboot = false,
   }) async {
     final int now = DateTime.now().millisecondsSinceEpoch;
     final int first = now + delay.inMilliseconds;
@@ -120,6 +125,7 @@ class AndroidAlarmManager {
       exact,
       wakeup,
       first,
+      rescheduleOnReboot,
       handle.toRawHandle(),
     ]);
     return (r == null) ? false : r;
@@ -145,6 +151,10 @@ class AndroidAlarmManager {
   /// alarm fires. If `wakeup` is false (the default), the device will not be
   /// woken up to service the alarm.
   ///
+  /// If `rescheduleOnReboot` is passed as `true`, the alarm will be persisted
+  /// across reboots. If `rescheduleOnReboot` is false (the default), the alarm
+  /// will not be rescheduled after a reboot and will not be executed.
+  ///
   /// Returns a [Future] that resolves to `true` on success and `false` on
   /// failure.
   static Future<bool> periodic(
@@ -153,6 +163,7 @@ class AndroidAlarmManager {
     dynamic Function() callback, {
     bool exact = false,
     bool wakeup = false,
+    bool rescheduleOnReboot = false,
   }) async {
     final int now = DateTime.now().millisecondsSinceEpoch;
     final int period = duration.inMilliseconds;
@@ -164,8 +175,15 @@ class AndroidAlarmManager {
     // TODO(amirh): remove this on when the invokeMethod update makes it to stable Flutter.
     // https://github.com/flutter/flutter/issues/26431
     // ignore: strong_mode_implicit_dynamic_method
-    final dynamic r = await _channel.invokeMethod('Alarm.periodic',
-        <dynamic>[id, exact, wakeup, first, period, handle.toRawHandle()]);
+    final dynamic r = await _channel.invokeMethod('Alarm.periodic', <dynamic>[
+      id,
+      exact,
+      wakeup,
+      first,
+      period,
+      rescheduleOnReboot,
+      handle.toRawHandle()
+    ]);
     return (r == null) ? false : r;
   }
 
