@@ -6,28 +6,22 @@
 
 #import <Firebase/Firebase.h>
 
-@interface FLTFirebaseDatabasePlugin ()
-@end
-
-@interface NSError (FlutterError)
-@property(readonly, nonatomic) FlutterError *flutterError;
-@property(readonly, nonatomic) NSDictionary *dictionary;
-@end
-
-@implementation NSError (FlutterError)
-- (FlutterError *)flutterError {
-  return [FlutterError errorWithCode:[NSString stringWithFormat:@"Error %ld", self.code]
-                             message:self.domain
-                             details:self.localizedDescription];
+static FlutterError *getFlutterError(NSError *error) {
+  return [FlutterError errorWithCode:[NSString stringWithFormat:@"Error %d", (int)error.code]
+                             message:error.domain
+                             details:error.localizedDescription];
 }
 
-- (NSDictionary *)dictionary {
+static NSDictionary *getDictionaryFromError(NSError *error) {
   return @{
-    @"code" : @(self.code),
-    @"message" : self.domain ?: [NSNull null],
-    @"details" : self.localizedDescription ?: [NSNull null],
+    @"code" : @(error.code),
+    @"message" : error.domain ?: [NSNull null],
+    @"details" : error.localizedDescription ?: [NSNull null],
   };
 }
+
+@interface FLTFirebaseDatabasePlugin ()
+
 @end
 
 FIRDatabaseReference *getReference(FIRDatabase *database, NSDictionary *arguments) {
@@ -173,7 +167,7 @@ id roundDoubles(id value) {
   }
   void (^defaultCompletionBlock)(NSError *, FIRDatabaseReference *) =
       ^(NSError *error, FIRDatabaseReference *ref) {
-        result(error.flutterError);
+        result(getFlutterError(error));
       };
   if ([@"FirebaseDatabase#goOnline" isEqualToString:call.method]) {
     [database goOnline];
@@ -280,7 +274,7 @@ id roundDoubles(id value) {
           // Invoke transaction completion on the Dart side.
           result(@{
             @"transactionKey" : call.arguments[@"transactionKey"],
-            @"error" : error.dictionary ?: [NSNull null],
+            @"error" : getDictionaryFromError(error) ?: [NSNull null],
             @"committed" : [NSNumber numberWithBool:committed],
             @"snapshot" : @{@"key" : snapshot.key ?: [NSNull null], @"value" : snapshot.value}
           });
@@ -314,7 +308,7 @@ id roundDoubles(id value) {
           [self.channel invokeMethod:@"Error"
                            arguments:@{
                              @"handle" : [NSNumber numberWithUnsignedInteger:handle],
-                             @"error" : error.dictionary,
+                             @"error" : getDictionaryFromError(error),
                            }];
         }];
     result([NSNumber numberWithUnsignedInteger:handle]);
