@@ -1,6 +1,10 @@
 package io.flutter.plugins.firebasemlvision;
 
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.graphics.Matrix;
 import android.net.Uri;
+import androidx.exifinterface.media.ExifInterface;
 import com.google.firebase.ml.vision.common.FirebaseVisionImage;
 import com.google.firebase.ml.vision.common.FirebaseVisionImageMetadata;
 import io.flutter.plugin.common.MethodCall;
@@ -44,7 +48,7 @@ public class FirebaseMlVisionPlugin implements MethodCallHandler {
       case "BarcodeDetector#detectInImage":
         BarcodeDetector.instance.handleDetection(image, options, result);
         break;
-      case "FaceDetector#detectInImage":
+      case "FaceDetector#processImage":
         FaceDetector.instance.handleDetection(image, options, result);
         break;
       case "LabelDetector#detectInImage":
@@ -63,11 +67,26 @@ public class FirebaseMlVisionPlugin implements MethodCallHandler {
 
   private FirebaseVisionImage dataToVisionImage(Map<String, Object> imageData) throws IOException {
     String imageType = (String) imageData.get("type");
+    assert imageType != null;
 
     switch (imageType) {
       case "file":
-        File file = new File((String) imageData.get("path"));
-        return FirebaseVisionImage.fromFilePath(registrar.context(), Uri.fromFile(file));
+        final String imageFilePath = (String) imageData.get("path");
+        final int rotation = getImageExifOrientation(imageFilePath);
+
+        if (rotation == 0) {
+          File file = new File(imageFilePath);
+          return FirebaseVisionImage.fromFilePath(registrar.context(), Uri.fromFile(file));
+        }
+
+        Matrix matrix = new Matrix();
+        matrix.postRotate(rotation);
+
+        final Bitmap bitmap = BitmapFactory.decodeFile(imageFilePath);
+        final Bitmap rotatedBitmap =
+            Bitmap.createBitmap(bitmap, 0, 0, bitmap.getWidth(), bitmap.getHeight(), matrix, true);
+
+        return FirebaseVisionImage.fromBitmap(rotatedBitmap);
       case "bytes":
         @SuppressWarnings("unchecked")
         Map<String, Object> metadataData = (Map<String, Object>) imageData.get("metadata");
@@ -80,9 +99,29 @@ public class FirebaseMlVisionPlugin implements MethodCallHandler {
                 .setRotation(getRotation((int) metadataData.get("rotation")))
                 .build();
 
-        return FirebaseVisionImage.fromByteArray((byte[]) imageData.get("bytes"), metadata);
+        byte[] bytes = (byte[]) imageData.get("bytes");
+        assert bytes != null;
+
+        return FirebaseVisionImage.fromByteArray(bytes, metadata);
       default:
         throw new IllegalArgumentException(String.format("No image type for: %s", imageType));
+    }
+  }
+
+  private int getImageExifOrientation(String imageFilePath) throws IOException {
+    ExifInterface exif = new ExifInterface(imageFilePath);
+    int orientation =
+        exif.getAttributeInt(ExifInterface.TAG_ORIENTATION, ExifInterface.ORIENTATION_NORMAL);
+
+    switch (orientation) {
+      case ExifInterface.ORIENTATION_ROTATE_90:
+        return 90;
+      case ExifInterface.ORIENTATION_ROTATE_180:
+        return 180;
+      case ExifInterface.ORIENTATION_ROTATE_270:
+        return 270;
+      default:
+        return 0;
     }
   }
 
