@@ -29,10 +29,10 @@ static FIRVisionFaceDetector *faceDetector;
                 face.hasRightEyeOpenProbability ? @(face.rightEyeOpenProbability) : [NSNull null];
 
             NSDictionary *data = @{
-              @"left" : @((int)face.frame.origin.x),
-              @"top" : @((int)face.frame.origin.y),
-              @"width" : @((int)face.frame.size.width),
-              @"height" : @((int)face.frame.size.height),
+              @"left" : @(face.frame.origin.x),
+              @"top" : @(face.frame.origin.y),
+              @"width" : @(face.frame.size.width),
+              @"height" : @(face.frame.size.height),
               @"headEulerAngleY" : face.hasHeadEulerAngleY ? @(face.headEulerAngleY)
                                                            : [NSNull null],
               @"headEulerAngleZ" : face.hasHeadEulerAngleZ ? @(face.headEulerAngleZ)
