@@ -9,14 +9,19 @@ part of firebase_ml_vision;
 /// A text recognizer is created via `textRecognizer()` in [FirebaseVision]:
 ///
 /// ```dart
-/// TextRecognizer textRecognizer = FirebaseVision.instance.textRecognizer();
+/// final FirebaseVisionImage image =
+///     FirebaseVisionImage.fromFilePath('path/to/file');
+///
+/// final TextRecognizer textRecognizer =
+///     FirebaseVision.instance.textRecognizer();
+///
+/// final List<VisionText> recognizedText =
+///     await textRecognizer.processImage(image);
 /// ```
-class TextRecognizer implements FirebaseVisionDetector {
+class TextRecognizer {
   TextRecognizer._();
 
   /// Detects [VisionText] from a [FirebaseVisionImage].
-  ///
-  /// The OCR is performed asynchronously.
   Future<VisionText> processImage(FirebaseVisionImage visionImage) async {
     final Map<dynamic, dynamic> reply =
         // TODO(amirh): remove this on when the invokeMethod update makes it to stable Flutter.
@@ -30,15 +35,6 @@ class TextRecognizer implements FirebaseVisionDetector {
     );
 
     return VisionText._(reply);
-  }
-
-  /// Detects [VisionText] from a [FirebaseVisionImage].
-  ///
-  /// The OCR is performed asynchronously.
-  @Deprecated('Please use `processImage`')
-  @override
-  Future<VisionText> detectInImage(FirebaseVisionImage visionImage) async {
-    return processImage(visionImage);
   }
 }
 
@@ -69,7 +65,7 @@ class RecognizedLanguage {
 abstract class TextContainer {
   TextContainer._(Map<dynamic, dynamic> data)
       : boundingBox = data['left'] != null
-            ? Rectangle<int>(
+            ? Rect.fromLTWH(
                 data['left'],
                 data['top'],
                 data['width'],
@@ -77,8 +73,8 @@ abstract class TextContainer {
               )
             : null,
         confidence = data['confidence'],
-        cornerPoints = List<Point<int>>.unmodifiable(
-            data['points'].map<Point<int>>((dynamic point) => Point<int>(
+        cornerPoints = List<Offset>.unmodifiable(
+            data['points'].map<Offset>((dynamic point) => Offset(
                   point[0],
                   point[1],
                 ))),
@@ -92,7 +88,7 @@ abstract class TextContainer {
   /// The point (0, 0) is defined as the upper-left corner of the image.
   ///
   /// Could be null even if text is found.
-  final Rectangle<int> boundingBox;
+  final Rect boundingBox;
 
   /// The confidence of the recognized text block.
   ///
@@ -106,7 +102,7 @@ abstract class TextContainer {
   /// rectangle. Parts of the region could be outside of the image.
   ///
   /// Could be empty even if text is found.
-  final List<Point<int>> cornerPoints;
+  final List<Offset> cornerPoints;
 
   /// All detected languages from recognized text.
   ///
