@@ -6,17 +6,11 @@
 
 #import <Firebase/Firebase.h>
 
-@interface NSError (FlutterError)
-@property(readonly, nonatomic) FlutterError *flutterError;
-@end
-
-@implementation NSError (FlutterError)
-- (FlutterError *)flutterError {
-  return [FlutterError errorWithCode:[NSString stringWithFormat:@"Error %ld", (long)self.code]
-                             message:self.domain
-                             details:self.localizedDescription];
+static FlutterError *getFlutterError(NSError *error) {
+  return [FlutterError errorWithCode:[NSString stringWithFormat:@"Error %d", (int)error.code]
+                             message:error.domain
+                             details:error.localizedDescription];
 }
-@end
 
 @interface FLTFirebaseStoragePlugin ()
 @property(nonatomic, retain) FlutterMethodChannel *channel;
@@ -303,7 +297,7 @@ typedef NS_ENUM(NSUInteger, StorageTaskEventType) {
   [ref dataWithMaxSize:[maxSize longLongValue]
             completion:^(NSData *_Nullable data, NSError *_Nullable error) {
               if (error != nil) {
-                result(error.flutterError);
+                result(getFlutterError(error));
                 return;
               }
               if (data == nil) {
@@ -330,7 +324,7 @@ typedef NS_ENUM(NSUInteger, StorageTaskEventType) {
   [task observeStatus:FIRStorageTaskStatusFailure
               handler:^(FIRStorageTaskSnapshot *snapshot) {
                 if (snapshot.error != nil) {
-                  result(snapshot.error.flutterError);
+                  result(getFlutterError(snapshot.error));
                 }
               }];
 }
@@ -340,7 +334,7 @@ typedef NS_ENUM(NSUInteger, StorageTaskEventType) {
   FIRStorageReference *ref = [storage.reference child:path];
   [ref metadataWithCompletion:^(FIRStorageMetadata *metadata, NSError *error) {
     if (error != nil) {
-      result(error.flutterError);
+      result(getFlutterError(error));
     } else {
       result([self buildDictionaryFromMetadata:metadata]);
     }
@@ -354,7 +348,7 @@ typedef NS_ENUM(NSUInteger, StorageTaskEventType) {
   [ref updateMetadata:[self buildMetadataFromDictionary:metadataDictionary]
            completion:^(FIRStorageMetadata *metadata, NSError *error) {
              if (error != nil) {
-               result(error.flutterError);
+               result(getFlutterError(error));
              } else {
                result([self buildDictionaryFromMetadata:metadata]);
              }
@@ -384,7 +378,7 @@ typedef NS_ENUM(NSUInteger, StorageTaskEventType) {
   FIRStorageReference *ref = [storage.reference child:path];
   [ref downloadURLWithCompletion:^(NSURL *URL, NSError *error) {
     if (error != nil) {
-      result(error.flutterError);
+      result(getFlutterError(error));
     } else {
       result(URL.absoluteString);
     }
@@ -396,7 +390,7 @@ typedef NS_ENUM(NSUInteger, StorageTaskEventType) {
   FIRStorageReference *ref = [storage.reference child:path];
   [ref deleteWithCompletion:^(NSError *error) {
     if (error != nil) {
-      result(error.flutterError);
+      result(getFlutterError(error));
     } else {
       result(nil);
     }
