@@ -4,11 +4,21 @@
 
 import 'dart:ui' show hashValues;
 import 'package:flutter/foundation.dart';
+import 'package:json_annotation/json_annotation.dart';
 import 'billing_client_wrapper.dart';
+import 'enum_converters.dart';
+import 'package:in_app_purchase/src/in_app_purchase_connection/product_details.dart';
+
+// WARNING: Changes to `@JsonSerializable` classes need to be reflected in the
+// below generated file. Run `flutter packages pub run build_runner watch` to
+// rebuild and watch for further changes.
+part 'sku_details_wrapper.g.dart';
 
 /// Dart wrapper around [`com.android.billingclient.api.SkuDetails`](https://developer.android.com/reference/com/android/billingclient/api/SkuDetails).
 ///
 /// Contains the details of an available product in Google Play Billing.
+@JsonSerializable()
+@SkuTypeConverter()
 class SkuDetailsWrapper {
   @visibleForTesting
   SkuDetailsWrapper({
@@ -33,28 +43,15 @@ class SkuDetailsWrapper {
   /// The map needs to have named string keys with values matching the names and
   /// types of all of the members on this class.
   @visibleForTesting
-  SkuDetailsWrapper.fromMap(Map<dynamic, dynamic> map)
-      : description = map['description'],
-        freeTrialPeriod = map['freeTrialPeriod'],
-        introductoryPrice = map['introductoryPrice'],
-        introductoryPriceMicros = map['introductoryPriceMicros'],
-        introductoryPriceCycles = map['introductoryPriceCycles'],
-        introductoryPricePeriod = map['introductoryPricePeriod'],
-        price = map['price'],
-        priceAmountMicros = map['priceAmountMicros'],
-        priceCurrencyCode = map['priceCurrencyCode'],
-        sku = map['sku'],
-        subscriptionPeriod = map['subscriptionPeriod'],
-        title = map['title'],
-        type = SkuType.fromString(map['type']),
-        isRewarded = map['isRewarded'];
+  factory SkuDetailsWrapper.fromJson(Map map) =>
+      _$SkuDetailsWrapperFromJson(map);
 
   final String description;
 
   /// Trial period in ISO 8601 format.
   final String freeTrialPeriod;
 
-  /// Introductory price, only applies to [SkuType.SUBS]. Formatted ("$0.99").
+  /// Introductory price, only applies to [SkuType.subs]. Formatted ("$0.99").
   final String introductoryPrice;
 
   /// [introductoryPrice] in micro-units 990000
@@ -78,7 +75,7 @@ class SkuDetailsWrapper {
   /// The product ID in Google Play Console.
   final String sku;
 
-  /// Applies to [SkuType.SUBS], formatted in ISO 8601.
+  /// Applies to [SkuType.subs], formatted in ISO 8601.
   final String subscriptionPeriod;
   final String title;
 
@@ -128,11 +125,23 @@ class SkuDetailsWrapper {
         type.hashCode,
         isRewarded.hashCode);
   }
+
+  /// Method to convert to the wrapper to the consolidated [ProductDetails] class.
+  ProductDetails toProductDetails() {
+    return ProductDetails(
+      id: sku,
+      title: title,
+      description: description,
+      price: price,
+    );
+  }
 }
 
 /// Translation of [`com.android.billingclient.api.SkuDetailsResponseListener`](https://developer.android.com/reference/com/android/billingclient/api/SkuDetailsResponseListener.html).
 ///
 /// Returned by [BillingClient.querySkuDetails].
+@JsonSerializable()
+@BillingResponseConverter()
 class SkuDetailsResponseWrapper {
   @visibleForTesting
   SkuDetailsResponseWrapper({@required this.responseCode, this.skuDetailsList});
@@ -141,17 +150,27 @@ class SkuDetailsResponseWrapper {
   ///
   /// The map needs to have named string keys with values matching the names and
   /// types of all of the members on this class.
-  SkuDetailsResponseWrapper.fromMap(Map<dynamic, dynamic> map)
-      : responseCode = BillingResponse.fromInt(map['responseCode']),
-        skuDetailsList =
-            List.castFrom<dynamic, Map<dynamic, dynamic>>(map['skuDetailsList'])
-                .map((Map<dynamic, dynamic> entry) =>
-                    SkuDetailsWrapper.fromMap(entry))
-                .toList();
+  factory SkuDetailsResponseWrapper.fromJson(Map<String, dynamic> map) =>
+      _$SkuDetailsResponseWrapperFromJson(map);
 
   /// The final status of the [BillingClient.querySkuDetails] call.
   final BillingResponse responseCode;
 
   /// A list of [SkuDetailsWrapper] matching the query to [BillingClient.querySkuDetails].
   final List<SkuDetailsWrapper> skuDetailsList;
+
+  @override
+  bool operator ==(dynamic other) {
+    if (other.runtimeType != runtimeType) {
+      return false;
+    }
+
+    final SkuDetailsResponseWrapper typedOther = other;
+    return typedOther is SkuDetailsResponseWrapper &&
+        typedOther.responseCode == responseCode &&
+        typedOther.skuDetailsList == skuDetailsList;
+  }
+
+  @override
+  int get hashCode => hashValues(responseCode, skuDetailsList);
 }
