@@ -6,6 +6,7 @@ import 'package:test/test.dart';
 import 'package:flutter/services.dart';
 
 import 'package:in_app_purchase/billing_client_wrappers.dart';
+import 'package:in_app_purchase/src/billing_client_wrappers/enum_converters.dart';
 import 'package:in_app_purchase/src/channel.dart';
 import '../stub_in_app_purchase_platform.dart';
 import 'sku_details_wrapper_test.dart';
@@ -37,18 +38,19 @@ void main() {
     test('returns BillingResponse', () async {
       stubPlatform.addResponse(
           name: 'BillingClient#startConnection(BillingClientStateListener)',
-          value: int.parse(BillingResponse.OK.toString()));
+          value: BillingResponseConverter().toJson(BillingResponse.ok));
       expect(
           await billingClient.startConnection(
               onBillingServiceDisconnected: () {}),
-          equals(BillingResponse.OK));
+          equals(BillingResponse.ok));
     });
 
     test('passes handle to onBillingServiceDisconnected', () async {
       final String methodName =
           'BillingClient#startConnection(BillingClientStateListener)';
       stubPlatform.addResponse(
-          name: methodName, value: int.parse(BillingResponse.OK.toString()));
+          name: methodName,
+          value: BillingResponseConverter().toJson(BillingResponse.ok));
       await billingClient.startConnection(onBillingServiceDisconnected: () {});
       final MethodCall call = stubPlatform.previousCallMatching(methodName);
       expect(call.arguments, equals(<dynamic, dynamic>{'handle': 0}));
@@ -68,33 +70,73 @@ void main() {
         'BillingClient#querySkuDetailsAsync(SkuDetailsParams, SkuDetailsResponseListener)';
 
     test('handles empty skuDetails', () async {
-      final BillingResponse responseCode = BillingResponse.DEVELOPER_ERROR;
+      final BillingResponse responseCode = BillingResponse.developerError;
       stubPlatform.addResponse(name: queryMethodName, value: <dynamic, dynamic>{
-        'responseCode': int.parse(responseCode.toString()),
+        'responseCode': BillingResponseConverter().toJson(responseCode),
         'skuDetailsList': <Map<String, dynamic>>[]
       });
 
       final SkuDetailsResponseWrapper response = await billingClient
           .querySkuDetails(
-              skuType: SkuType.INAPP, skusList: <String>['invalid']);
+              skuType: SkuType.inapp, skusList: <String>['invalid']);
 
       expect(response.responseCode, equals(responseCode));
       expect(response.skuDetailsList, isEmpty);
     });
 
     test('returns SkuDetailsResponseWrapper', () async {
-      final BillingResponse responseCode = BillingResponse.OK;
+      final BillingResponse responseCode = BillingResponse.ok;
       stubPlatform.addResponse(name: queryMethodName, value: <String, dynamic>{
-        'responseCode': int.parse(responseCode.toString()),
+        'responseCode': BillingResponseConverter().toJson(responseCode),
         'skuDetailsList': <Map<String, dynamic>>[buildSkuMap(dummyWrapper)]
       });
 
       final SkuDetailsResponseWrapper response = await billingClient
           .querySkuDetails(
-              skuType: SkuType.INAPP, skusList: <String>['invalid']);
+              skuType: SkuType.inapp, skusList: <String>['invalid']);
 
       expect(response.responseCode, equals(responseCode));
       expect(response.skuDetailsList, contains(dummyWrapper));
+    });
+  });
+
+  group('launchBillingFlow', () {
+    final String launchMethodName =
+        'BillingClient#launchBillingFlow(Activity, BillingFlowParams)';
+
+    test('serializes and deserializes data', () async {
+      final BillingResponse sentCode = BillingResponse.ok;
+      stubPlatform.addResponse(
+          name: launchMethodName,
+          value: BillingResponseConverter().toJson(sentCode));
+      final SkuDetailsWrapper skuDetails = dummyWrapper;
+      final String accountId = "hashedAccountId";
+
+      final BillingResponse receivedCode = await billingClient
+          .launchBillingFlow(skuDetails: skuDetails, accountId: accountId);
+
+      expect(receivedCode, equals(sentCode));
+      Map<dynamic, dynamic> arguments =
+          stubPlatform.previousCallMatching(launchMethodName).arguments;
+      expect(arguments['sku'], equals(skuDetails.sku));
+      expect(arguments['accountId'], equals(accountId));
+    });
+
+    test('handles null accountId', () async {
+      final BillingResponse sentCode = BillingResponse.ok;
+      stubPlatform.addResponse(
+          name: launchMethodName,
+          value: BillingResponseConverter().toJson(sentCode));
+      final SkuDetailsWrapper skuDetails = dummyWrapper;
+
+      final BillingResponse receivedCode =
+          await billingClient.launchBillingFlow(skuDetails: skuDetails);
+
+      expect(receivedCode, equals(sentCode));
+      Map<dynamic, dynamic> arguments =
+          stubPlatform.previousCallMatching(launchMethodName).arguments;
+      expect(arguments['sku'], equals(skuDetails.sku));
+      expect(arguments['accountId'], isNull);
     });
   });
 }
