@@ -6,9 +6,12 @@ import 'package:test/test.dart';
 
 import 'package:flutter/widgets.dart';
 import 'package:in_app_purchase/billing_client_wrappers.dart';
+import 'package:in_app_purchase/src/billing_client_wrappers/enum_converters.dart';
 import 'package:in_app_purchase/src/in_app_purchase_connection/google_play_connection.dart';
 import 'package:in_app_purchase/src/channel.dart';
 import '../stub_in_app_purchase_platform.dart';
+import 'package:in_app_purchase/src/in_app_purchase_connection/product_details.dart';
+import '../billing_client_wrappers/sku_details_wrapper_test.dart';
 
 void main() {
   final StubInAppPurchasePlatform stubPlatform = StubInAppPurchasePlatform();
@@ -24,7 +27,7 @@ void main() {
     WidgetsFlutterBinding.ensureInitialized();
     stubPlatform.addResponse(
         name: startConnectionCall,
-        value: int.parse(BillingResponse.OK.toString()));
+        value: BillingResponseConverter().toJson(BillingResponse.ok));
     stubPlatform.addResponse(name: endConnectionCall, value: null);
     connection = GooglePlayConnection.instance;
   });
@@ -61,6 +64,52 @@ void main() {
     test('false', () async {
       stubPlatform.addResponse(name: 'BillingClient#isReady()', value: false);
       expect(await connection.isAvailable(), isFalse);
+    });
+  });
+
+  group('querySkuDetails', () {
+    final String queryMethodName =
+        'BillingClient#querySkuDetailsAsync(SkuDetailsParams, SkuDetailsResponseListener)';
+
+    test('handles empty skuDetails', () async {
+      final BillingResponse responseCode = BillingResponse.developerError;
+      stubPlatform.addResponse(name: queryMethodName, value: <dynamic, dynamic>{
+        'responseCode': BillingResponseConverter().toJson(responseCode),
+        'skuDetailsList': <Map<String, dynamic>>[]
+      });
+
+      final ProductDetailsResponse response =
+          await connection.queryProductDetails(<String>[''].toSet());
+      expect(response.productDetails, isEmpty);
+    });
+
+    test('should get correct product details', () async {
+      final BillingResponse responseCode = BillingResponse.ok;
+      stubPlatform.addResponse(name: queryMethodName, value: <String, dynamic>{
+        'responseCode': BillingResponseConverter().toJson(responseCode),
+        'skuDetailsList': <Map<String, dynamic>>[buildSkuMap(dummyWrapper)]
+      });
+      // Since queryProductDetails makes 2 platform method calls (one for each SkuType), the result will contain 2 dummyWrapper instead
+      // of 1.
+      final ProductDetailsResponse response =
+          await connection.queryProductDetails(<String>['valid'].toSet());
+      expect(response.productDetails.first.title, dummyWrapper.title);
+      expect(
+          response.productDetails.first.description, dummyWrapper.description);
+      expect(response.productDetails.first.price, dummyWrapper.price);
+    });
+
+    test('should get the correct notFoundIDs', () async {
+      final BillingResponse responseCode = BillingResponse.ok;
+      stubPlatform.addResponse(name: queryMethodName, value: <String, dynamic>{
+        'responseCode': BillingResponseConverter().toJson(responseCode),
+        'skuDetailsList': <Map<String, dynamic>>[buildSkuMap(dummyWrapper)]
+      });
+      // Since queryProductDetails makes 2 platform method calls (one for each SkuType), the result will contain 2 dummyWrapper instead
+      // of 1.
+      final ProductDetailsResponse response =
+          await connection.queryProductDetails(<String>['invalid'].toSet());
+      expect(response.notFoundIDs.first, 'invalid');
     });
   });
 }
